Refactor CacheFS for efficiency and usability.
The main class is now CacheFSMixin, a mixin class that can add caching to
any FS implementation.  It operates explicitly on a PathMap of recent 
info dicts, making it more efficient and more robust than the old CacheFS.

The CacheFS wrapper class is still there, but it's a trivial class that
just mixes CacheFSMixin into the base WrapFS class.

TahoeFS has also been modified to use the mixin architecture.


git-svn-id: http://pyfilesystem.googlecode.com/svn/trunk@610 67cdc799-7952-0410-af00-57a81ceafa0f